Moving to deploybot
We will use deploybot for the distributed
caching project after all.  This means we
will have a different pattern for passing
variables and files around.

Also, we don't want to write our own tests
for something we aren't modifying and that
has passing tests upstream.  So, we are
going to have a dummy test that should
always pass to make circle happy and not
notify us in Slack that we have problems
due to how our Slack integrations work with
circle.