+build number
So if you don't refer to the build number
in the tests, it will attempt to use "latest"
which we don't have (and don't want to use anyway
for new builds.) So even in the tests you need
to specify which *exact* image to use, which in this
case means using the $CIRCLE_BUILD_NUMBER variable.
I should've thought of that. :)